Setup: refactoring of tests
